requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
python-dotenv==1.0.0
PyMuPDF==1.23.8
python-docx==1.1.0
emoji==2.10.0
Pillow==10.2.0
transformers==4.36.0
tiktoken==0.5.2

# Advanced Security Dependencies
numpy==1.24.3
scipy==1.11.4
scikit-image==0.22.0
opencv-python==4.8.1.78
pycryptodome==3.19.0
unidecode==1.3.7
charset-normalizer==3.3.2
fastapi-limiter==0.1.5
aioredis==2.0.1
async-timeout==4.0.3

